requirements.txt
# Copyright (c) 2019-2021 The University of Manchester
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.

SpiNNUtilities == 1!6.0.1
SpiNNMachine == 1!6.0.1
SpiNNMan == 1!6.0.1
SpiNNaker_PACMAN == 1!6.0.1
SpiNNaker_DataSpecification == 1!6.0.1
SpiNNFrontEndCommon == 1!6.0.1
sPyNNaker == 1!6.0.1
numpy == 1.19; python_version == '3.6'
numpy >= 1.19, <= 1.20; python_version == '3.7'
numpy; python_version >= '3.8'
lxml
opencv-python == 4.5.3.56
